requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2023.7.22
charset-normalizer==3.3.0
idna==3.4
python-telegram-bot==13.14
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tornado==6.1
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.6
